Lincoln built a train station by applying parallel programming and concurrency principles.
Assignment 1

Write a control panel GUI for controlling a train station, with the following rules; or any variant on them that you ask me and I approve:
There is at least one round-trip destination.
There is at least one track.
There is at least one train with at least one seat.
Ticketed passengers arrive at a platform, waiting until a seat is available to enter.
All passengers leave the train when it arrives at the platform.
A train leaves when the track is clear and either all seats are occupied, or when the operator clicks "start".
Clicking start when trains are in motion has no effect.
In order to test it, you will have to create simulated Passengers that randomly arrive at some average rate, and a simulated run that pretends to go from here to there for a while.
All details of the GUI are up to you, but you should minimally indicate the state of the train(s), the number of passengers, the number waiting, and average throughput.